Change Session code

+++ src/main/java/com/homework/session/service/LoginService.java
@@ -4,12 +4,17 @@ import com.homework.session.Repository.UserRepository;
 import com.homework.session.dto.UserDto;
 import com.homework.session.entity.User;
 import com.homework.session.error.exception.UnAuthorizedException;
+import com.homework.session.sessionManager.SessionManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import static com.homework.session.error.ErrorCode.ACCESS_DENIED_EXCEPTION;
 
@@ -20,15 +25,23 @@ import static com.homework.session.error.ErrorCode.ACCESS_DENIED_EXCEPTION;
 public class LoginService {
 
     private final UserRepository userRepository;
+    private final SessionManager sessionManager;
 
     @Autowired
     private PasswordEncoder passwordEncoder;
 
     @Transactional
-    public void login(String email, String password) {
+    public void login(String email, String password, HttpServletResponse response) {
         userRepository.findByEmail(email)
-                .filter(u -> passwordEncoder.matches(u.getPassword(), password))
-                .orElse(null );
+                .filter(u -> passwordEncoder.matches(password, u.getPassword()))
+                .orElseThrow(() -> { throw new UnAuthorizedException("E0002", ACCESS_DENIED_EXCEPTION); });
+
+        UserDto userDto = UserDto.builder()
+                .email(email)
+                .password(passwordEncoder.encode(password))
+                .build();
+
+        sessionManager.createSession(userDto, response);
     }
 
     @Transactional
@@ -38,35 +51,44 @@ public class LoginService {
         }
 
         User user = User.builder()
-                        .email(userDto.getEmail())
-                        .password(passwordEncoder.encode(userDto.getPassword()))
-                        .phoneNumber(userDto.getPhoneNumber())
-                        .build();
+                .email(userDto.getEmail())
+                .nickName(userDto.getNickName())
+                .password(passwordEncoder.encode(userDto.getPassword()))
+                .phoneNumber(userDto.getPhoneNumber())
+                .build();
 
         userRepository.save(user);
     }
 
     @Transactional
-    public void update(UserDto userDto) {
+    public void update(UserDto userDto, HttpServletRequest request) {
 
         User user = userRepository.findByEmail(userDto.getEmail()).orElseThrow(() ->
                     { throw new UnAuthorizedException("E0002", ACCESS_DENIED_EXCEPTION); });
 
         UserDto updateDto = UserDto.builder()
                 .email(userDto.getEmail())
+                .nickName(userDto.getNickName())
                 .password(passwordEncoder.encode(userDto.getPassword()))
                 .phoneNumber(userDto.getPhoneNumber())
                 .build();
 
-        user.update(updateDto);
+        HttpSession session = request.getSession();
+
+        if (session != null) {
+            user.update(updateDto);
+        }
     }
 
     @Transactional
-    public void delete(UserDto userDto) {
+    public void delete(UserDto userDto, HttpServletRequest request) {
         User user = userRepository.findByEmail(userDto.getEmail()).orElseThrow(() ->
             { throw new UnAuthorizedException("E0002", ACCESS_DENIED_EXCEPTION); });
 
-        log.info("탈퇴처리 되었습니다.");
-        userRepository.delete(user);
+        HttpSession session = request.getSession();
+
+        if (session != null) {
+            userRepository.delete(user);
+        }
     }
 }
